@@ -9,19 +9,19 @@
     "preview": "vite preview"
   },
   "dependencies": {
-    "@chakra-ui/react": "^1.8.6",
+    "@chakra-ui/react": "^2",
     "@emotion/react": "^11",
     "@emotion/styled": "^11",
-    "@lego/klik-ui": "^0.56.4",
-    "framer-motion": "^6",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2"
+    "@lego/klik-ui": "^1.0.0-canary.1373.3417539357.0",
+    "framer-motion": "^7",
+    "react": "^18",
+    "react-dom": "^18"
   },
   "devDependencies": {
-    "@types/react": "^17",
-    "@types/react-dom": "^17",
+    "@types/react": "^18",
+    "@types/react-dom": "^18",
     "@vitejs/plugin-react": "^2.0.1",
-    "typescript": "^4.6.4",
+    "typescript": "4.6.4",
     "vite": "^3.0.7"
   }
 }
