@@ -1,14 +1,21 @@
-import { ChakraProvider } from "@chakra-ui/react";
-import { KlikProvider } from "@lego/klik-ui";
+import { createStandaloneToast as createChakraStandaloneToast } from "@chakra-ui/react";
+import { createStandaloneToast, KlikProvider, theme } from "@lego/klik-ui";
 import React from "react";
-import ReactDOM from "react-dom";
+import { createRoot } from "react-dom/client";
 import App from "./App";
 
-ReactDOM.render(
+const { ToastContainer: KlikToastContainer } = createStandaloneToast();
+const { ToastContainer: ChakraToastContainer } = createChakraStandaloneToast();
+
+const container = document.getElementById("root");
+const root = createRoot(container!);
+
+root.render(
   <React.StrictMode>
-    <KlikProvider>
+    <KlikProvider theme={theme}>
       <App />
+      <KlikToastContainer />
+      <ChakraToastContainer />
     </KlikProvider>
-  </React.StrictMode>,
-  document.getElementById("root")
+  </React.StrictMode>
 );
